+++ src/dataTable/dataTable.ts
@@ -67,7 +67,7 @@ export class DataTable<T> implements IDataTable<T> {
 
         await this.loadConfig();
 
-        let dataTableIndex;
+        let dataTableIndex: DataTableIndex;
         if (!StringHelper.isEmpty(this._config.indexBundleHash)) {
             const indexBundleHash = Hash.fromTrytes(Trytes.fromString(this._config.indexBundleHash));
 
@@ -76,9 +76,17 @@ export class DataTable<T> implements IDataTable<T> {
             this.updateProgress(1, 1, "Retrieving Index");
 
             if (index && index.length > 0) {
-                const objectToTrytesConverter = new ObjectTrytesConverter<DataTableIndex>();
+                const objectToTrytesConverter = new ObjectTrytesConverter<DataTableIndex | string[]>();
 
-                dataTableIndex = objectToTrytesConverter.from(index[0].data);
+                const dt = objectToTrytesConverter.from(index[0].data);
+                if (ArrayHelper.isArray(dt)) {
+                    dataTableIndex = {
+                        bundles: <string[]>dt,
+                        lastIdx: Hash.EMPTY.toTrytes().toString()
+                    };
+                } else {
+                    dataTableIndex = <DataTableIndex>dt;
+                }
 
                 this._logger.info("<=== DataTable::index", dataTableIndex);
             } else {
@@ -99,7 +107,7 @@ export class DataTable<T> implements IDataTable<T> {
 
         await this.loadConfig();
 
-        await this.saveIndex([]);
+        await this.saveIndexBundleHashes([]);
     }
 
     /**
@@ -133,11 +141,10 @@ export class DataTable<T> implements IDataTable<T> {
 
         const addHash = storageItem.bundleHash.toTrytes().toString();
 
-        let index = await this.index();
-        index = index || [];
-        index.push(addHash);
+        const indexBundleHashes = await this.loadIndexBundleHashes();
+        indexBundleHashes.push(addHash);
 
-        await this.saveIndex(index);
+        await this.saveIndexBundleHashes(indexBundleHashes);
 
         this._logger.info("<=== DataTable::store", storageItem.bundleHash);
 
@@ -158,11 +165,11 @@ export class DataTable<T> implements IDataTable<T> {
 
         const hashes = [];
 
-        let index;
+        let indexBundleHashes;
         if (!clearIndex) {
-            index = await this.index();
+            indexBundleHashes = await this.loadIndexBundleHashes();
         }
-        index = index || [];
+        indexBundleHashes = indexBundleHashes || [];
 
         this.updateProgress(0, data.length, "Storing Items");
         for (let i = 0; i < data.length; i++) {
@@ -185,12 +192,12 @@ export class DataTable<T> implements IDataTable<T> {
 
             const addHash = storageItem.bundleHash.toTrytes().toString();
 
-            index.push(addHash);
+            indexBundleHashes.push(addHash);
 
             hashes.push(storageItem.bundleHash);
         }
 
-        await this.saveIndex(index);
+        await this.saveIndexBundleHashes(indexBundleHashes);
 
         this._logger.info("<=== DataTable::storeMultiple", hashes);
 
@@ -227,19 +234,18 @@ export class DataTable<T> implements IDataTable<T> {
             enumerable: true
         });
 
-        let index = await this.index();
-        index = index || [];
+        const indexBundleHashes = await this.loadIndexBundleHashes();
         const removeHash = originalId.toTrytes().toString();
         const addHash = storageItem.bundleHash.toTrytes().toString();
 
-        const idx = index.indexOf(removeHash);
+        const idx = indexBundleHashes.indexOf(removeHash);
         if (idx >= 0) {
-            index.splice(idx, 1, addHash);
+            indexBundleHashes.splice(idx, 1, addHash);
         } else {
-            index.push(addHash);
+            indexBundleHashes.push(addHash);
         }
 
-        await this.saveIndex(index);
+        await this.saveIndexBundleHashes(indexBundleHashes);
 
         this._logger.info("<=== DataTable::update", storageItem.bundleHash);
 
@@ -290,9 +296,9 @@ export class DataTable<T> implements IDataTable<T> {
         if (ArrayHelper.isTyped(ids, Hash)) {
             loadIds = ids;
         } else {
-            const index = await this.index();
-            if (ArrayHelper.isTyped(index, String)) {
-                loadIds = index.map(b => Hash.fromTrytes(Trytes.fromString(b)));
+            const indexBundleHashes = await this.loadIndexBundleHashes();
+            if (ArrayHelper.isTyped(indexBundleHashes, String)) {
+                loadIds = indexBundleHashes.map(b => Hash.fromTrytes(Trytes.fromString(b)));
             }
         }
 
@@ -336,15 +342,14 @@ export class DataTable<T> implements IDataTable<T> {
 
         await this.loadConfig();
 
-        let index = await this.index();
-        index = index || [];
+        const indexBundleHashes = await this.loadIndexBundleHashes();
         const removeHash = id.toTrytes().toString();
 
-        const idx = index.indexOf(removeHash);
+        const idx = indexBundleHashes.indexOf(removeHash);
         if (idx >= 0) {
-            index.splice(idx, 1);
+            indexBundleHashes.splice(idx, 1);
 
-            await this.saveIndex(index);
+            await this.saveIndexBundleHashes(indexBundleHashes);
 
             this._logger.info("<=== DataTable::remove");
         } else {
@@ -361,16 +366,15 @@ export class DataTable<T> implements IDataTable<T> {
 
         await this.loadConfig();
 
-        let index = await this.index();
-        index = index || [];
+        const indexBundleHashes = await this.loadIndexBundleHashes();
         let removed = false;
 
         for (let i = 0; i < ids.length; i++) {
             const removeHash = ids[i].toTrytes().toString();
 
-            const idx = index.indexOf(removeHash);
+            const idx = indexBundleHashes.indexOf(removeHash);
             if (idx >= 0) {
-                index.splice(idx, 1);
+                indexBundleHashes.splice(idx, 1);
 
                 this._logger.info("<=== DataTable::removeMultiple", ids[i]);
 
@@ -381,7 +385,7 @@ export class DataTable<T> implements IDataTable<T> {
         }
 
         if (removed) {
-            await this.saveIndex(index);
+            await this.saveIndexBundleHashes(indexBundleHashes);
         }
     }
 
@@ -417,12 +421,24 @@ export class DataTable<T> implements IDataTable<T> {
     }
 
     /* @internal */
-    private async saveIndex(index: DataTableIndex): Promise<void> {
+    private async loadIndexBundleHashes(): Promise<string[]> {
+        const idx = await this.index();
+
+        return idx && idx.bundles ? idx.bundles : [];
+    }
+
+    /* @internal */
+    private async saveIndexBundleHashes(bundles: string[]): Promise<void> {
         const indexAddress = Address.fromTrytes(Trytes.fromString(this._config.indexAddress));
+
+        const dataTableIndex = {
+            bundles,
+            lastIdx: this._config.indexBundleHash || Hash.EMPTY.toTrytes().toString()
+        };
 
         const objectToTrytesConverterIndex = new ObjectTrytesConverter<DataTableIndex>();
 
-        const trytesIndex = objectToTrytesConverterIndex.to(index);
+        const trytesIndex = objectToTrytesConverterIndex.to(dataTableIndex);
 
         this.updateProgress(0, 1, "Storing Index");
         const indexStorageItem = await this._storageClient.save(indexAddress, trytesIndex, DataTable.INDEX_TAG);