+++ src/interfaces/dataTableIndex.ts
@@ -1,4 +1,7 @@
 /**
  * Represents a table index for storing data.
  */
-export type DataTableIndex = string[];
+export type DataTableIndex = {
+    bundles: string[];
+    lastIdx: string;
+};