+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.9.5
+
+* Change DataTableIndex format, added lastIdx to facilitate audit trail
+* Fixed setting of unsafe headers on Amazon AWS signer
+
+## v0.9.4
+
+* Fixed repo links in readme and coverage config
+
 ## v0.9.3
 
 * First public release